move counter list panel components to pages/counter/components/
