+++ src/pages/counter/components/CounterReportsListPanel.vue
